Tarefa1_2022li1g029

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

A função mapaValido vai receber um tipo Mapa e vai dar um Booliano, esta função vai dizer se o mapa recebido é um mapa em que se possa jogar ou se é um mapa que não pode existir.Esta função também esta definida por outras funções auxiliares como mapaValido1 , mapaValido2, 'mapaValido3, 'mapaValido4 , 'mapaValido5 , 'mapaValido6 , mapaValido7 cada uma desta têm funções especificas que serão explicadas mais tarde.

Exemplo de Utilização:

>>> mapaValido (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Tronco]),(Rio (-1),[Nenhum,Nenhum,Tronco]),(Relva,[Nenhum,Nenhum,Arvore]),(Estrada 1,[Carro,Nenhum,Carro]),(Rio 1,[Tronco,Nenhum,Tronco])])
>>> True

>>> mapaValido (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Tronco]),(Rio 1,[Nenhum,Nenhum,Tronco]),(Relva,[Nenhum,Nenhum,Arvore]),(Estrada 1,[Carro,Carro,Carro]),(Rio 1,[Tronco,Nenhum,Tronco])])]
>>> False

mapaValido1 :: Mapa -> Bool

Como referi anteriormentente a função principal, mapaValido é composta por algumas auxiliares sendo uma delas a função mapaValido1, esta função recebe um mapa e confirma se nesse mapa consoante o terreno os obstaculos estão corretos.

Exemplo de Utilização:

>>> mapaValido1 (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Arvore])])
>>> False

>>> mapaValido1 (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Tronco])])
>>> True

>>> mapaValido1 (Mapa 3 [(Relva,[Nenhum,Nenhum,Carro]),(Estrada 3,[Nenhum,Nenhum,Carro]),(Estrada 3,[Nenhum,Nenhum,Arvore])])
>>> False

mapaValido2 :: Mapa -> Bool

A função mapaValido2 recebe um Mapa e compara os terrenos, se o terrenos forem Estrada , Relva , Rio e depois outra qualquer a função faz a recursiva se os terrenos forem Rio e depois outro Rio a função compara as velocidades. Se ambas forem tiverem o mesmo sinal então ambos os rios estão a dirigir se para a mesma direção. Esta função pretende que os seguidos nunca estejam a ir pela mesma direção.

Exemplos de Utilização:

>>> mapaValido2 (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Tronco]),(Rio (-1),[Nenhum,Nenhum,Tronco]),(Relva,[Nenhum,Nenhum,Arvore]),(Rio (-1),[Nenhum,Nenhum,Carro]),(Rio 1,[Nenhum,Nenhum,Carro])])
>>> True

>>> mapaValido2 (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Tronco]),(Rio (-1),[Nenhum,Nenhum,Carro]),(Relva,[Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Carro]),(Rio 1,[Nenhum,Nenhum,Carro])])
>>> False

mapaValido3 :: Mapa -> Bool

A função mapaValido3 tem como objetivo contar a quantidade de troncos que existe num rio e caso esse valor for maior que 5 então a função diz que Mapa não é valido , para isso é utilizado uma função auxiliar chamada aux3 que junta os elementos iguais da lista de obstaculos em listas, assim é só pedir o comprimento da lista se a cabeça da lista de obstaculos for Tronco. Nesta função existe uma exceção que caso na lista de obstaculos o primeiro elemento for tronco e igual ao ultimo então, ter se á de somar o comprimento do primeiro elemento da lista com o comprimento do ultimo elemento da lista dos auxiliares.

Exemplos de Utilização:

>>> mapaValido3 (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Tronco]),(Rio (-1),[Nenhum,Nenhum,Tronco]),(Relva,[Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Carro]),(Rio 1,[Tronco,Tronco,Tronco,Nenhum,Nenhum,Tronco,Tronco])])
>>> True

>>> mapaValido3 (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Tronco]),(Rio (-1),[Nenhum,Nenhum,Tronco]),(Relva,[Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Carro]),(Rio 1,[Tronco,Tronco,Tronco,Nenhum,Tronco,Tronco,Tronco])])
>>> False

aux3 :: Eq a => [a] -> [[a]]

Esta função aux3 simplesmente recebe a lista dos obstaculos e cria uma lista de listas com os elementos iguais dessa lista juntos Utilizando o exemplo da função mapaValido3.

Exemplos de Utilização:

>>> aux3 [Tronco,Tronco,Tronco,Nenhum,Nenhum,Tronco,Tronco]
>>> [[Tronco,Tronco,Tronco],[Nenhum,Nenhum],[Tronco,Tronco]]

>>> aux3 [Tronco,Tronco,Tronco,Nenhum,Nenhum]
>>> [[Tronco,Tronco,Tronco],[Nenhum,Nenhum]]

mapaValido4 :: Mapa -> Bool

A função mapaValido4 funciona da mesma maneira que a função mapaValido3 mas com Carro e com a auxiliar aux4 .

Exemplo de Utilização:

>>> mapaValido4 (Mapa 3 [(Estrada 3,[Nenhum,Nenhum,Carro]),(Rio (-1),[Nenhum,Nenhum,Tronco]),(Relva,[Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Carro]),(Rio 1,[Tronco,Tronco,Tronco,Nenhum,Nenhum,Tronco,Tronco])])
>>> True

aux4 :: Eq a => [a] -> [[a]]

A função auxiliar aux4 funciona da mesma maneira que a função aux3.

mapaValido5 :: Mapa -> Bool

A função mapaValido5 faz com que ,em todas as listas de obstaculos, exista pelo menos um elemento Nenhum, ou seja, uma linha não pode ser formada somente por obstáculos

Exemplo de Utilização:

>>> mapaValido5 (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Tronco]),(Rio (-1),[Nenhum,Nenhum,Tronco])
>>> True

>>> mapaValido5 (Mapa 3 [(Rio 3,[Tronco,Tronco,Tronco]),(Rio (-1),[Nenhum,Nenhum,Tronco])
>>> False

mapaValido6 :: Mapa -> Bool

A função mapaValido6 recebe um Mapa e o que faz é comparar a largura desse mapa com o comprimento da lista de obstaculos. Para que um mapa seja válido a largura e o comprimento da lista de obstáculos tem de ser a mesma.

Exemplos de Utilização:

>>> mapaValido6 (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Tronco]),(Rio (-1),[Nenhum,Nenhum,Tronco])])
>>> True

>>> mapaValido6 (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Tronco]),(Rio (-1),[Nenhum,Nenhum,Tronco,Tronco])])
>>> False

mapaValido7 :: Mapa -> Bool

A função mapaValido7 recebe um mapa e caso existam mais de 5 relvas ou estradas seguidas ou mais de 4 rios seguidos então o mapa não será valido

Exemplos de Utilização:

>>> mapaValido7 (Mapa 3 [(Rio 3,[Nenhum,Nenhum,Tronco]),(Rio (-1),[Nenhum,Nenhum,Tronco]),(Relva,[Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Carro])])
>>> True

>>> mapaValido7 (Mapa 3 [(Relva,[Nenhum,Nenhum,Tronco]),(Relva,[Nenhum,Nenhum,Tronco]),(Relva,[Nenhum,Nenhum,Arvore]),(Relva,[Nenhum,Nenhum,Carro]),(Relva,[Tronco,Tronco,Nenhum,Tronco,Tronco]),(Relva,[Tronco,Tronco,Nenhum,Tronco,Tronco])])
>>> False